+++ headers/winning_combinations.h
@@ -1,7 +1,7 @@
 #ifndef __WINNING_COMBINATIONS_H__
 #define __WINNING_COMBINATIONS_H__
 
-unsigned check_winning_combinations();
-unsigned check_column_full_error();
+unsigned short check_winning_combinations();
+unsigned short check_column_full_error();
 
 #endif
