+++ src/computer.c
@@ -10,21 +10,21 @@ Le score d'une colonne correspond au nombre de jetons de la couleur jouée par
 l'ordinateur si ce dernier décide de jouer dans cette colonne.
 Le score d'une colonne est forcément positif ou nul.
 */
-static unsigned moves_scores[7];
+static unsigned short moves_scores[7];
 
-void undo_move(unsigned grid[6][7], const unsigned column) {
-    unsigned row = 0;
+void undo_move(unsigned short grid[6][7], const unsigned short column) {
+    unsigned short row = 0;
     // Trouver la première ligne non-vide dans la colonne
     while (row < N_ROWS && grid[row][column] == NO_TOKEN)
         row++;
     grid[row][column] = NO_TOKEN;
 }
 
-unsigned check_winning_move(unsigned grid[6][7]) {
-    for (unsigned column = 0; column < N_COLUMNS; column++) {
+unsigned short check_winning_move(unsigned short grid[6][7]) {
+    for (unsigned short column = 0; column < N_COLUMNS; column++) {
         // Simuler un coup de l'ordinateur dans la colonne 'column'
         // L'ordinateur joue toujours un jeton rouge
-        unsigned result = play_token(grid, column, RED);
+        unsigned short result = play_token(grid, column, RED);
         if (result == NO_ERROR) {
             // Vérifier si ce coup mène à une victoire
             if (check_game_status(grid) != GAME_CONTINUE) {
@@ -46,27 +46,27 @@ unsigned check_winning_move(unsigned grid[6][7]) {
     return NO_WINNING_MOVE;
 }
 
-void compute_column_scores(unsigned grid[6][7]) {
+void compute_column_scores(unsigned short grid[6][7]) {
     /*
     Complète le tableau moves_scores en calculant le score de chaque colonne
     */
-   for (unsigned col = 0; col < N_COLUMNS; col++) {
-        unsigned result = play_token(grid, col, COMPUTER_TOKEN_COLOR);
+   for (unsigned short col = 0; col < N_COLUMNS; col++) {
+        unsigned short result = play_token(grid, col, COMPUTER_TOKEN_COLOR);
         if (result == COLUMN_FULL)
             moves_scores[col] = 0;
         else if (result == NO_ERROR) {
             // Trouver la première ligne non-vide dans la colonne
-            unsigned row = 0;
+            unsigned short row = 0;
             while (row < N_ROWS && grid[row][col] == NO_TOKEN)
                 row++;
             // Récupérer la couleur du jeton
-            unsigned token = grid[row][col];
-            unsigned start_row = row;
-            unsigned start_col = col;
+            unsigned short token = grid[row][col];
+            unsigned short start_row = row;
+            unsigned short start_col = col;
             // Calculer combien de pièces seraient alignées verticalement
-            unsigned vertical_count = 0;
-            unsigned horizontal_count = 0;
-            unsigned diagonal_count = 0;
+            unsigned short vertical_count = 0;
+            unsigned short horizontal_count = 0;
+            unsigned short diagonal_count = 0;
             // Compter le nombre de jetons de la même couleur sur la colonne
             while (row < N_ROWS && grid[row][col] == token) {
                 row++;
@@ -81,7 +81,7 @@ void compute_column_scores(unsigned grid[6][7]) {
                 horizontal_count++;
             }
             col = start_col - 1;
-            while (col != (unsigned) - 1 && grid[row][col] == token) {
+            while (col != (unsigned short) - 1 && grid[row][col] == token) {
                 col--;
                 horizontal_count++;
             }
@@ -97,31 +97,31 @@ void compute_column_scores(unsigned grid[6][7]) {
             }
             row = start_row - 1;
             col = start_col - 1;
-            while (row != (unsigned) - 1 && col != (unsigned) - 1 && grid[row][col] == token){
+            while (row != (unsigned short) - 1 && col != (unsigned short) - 1 && grid[row][col] == token){
                 row--;
                 col--;
                 diagonal_count++;
             }
             col = start_col;
             // Le score de la colonne est le max de horizontal_count, vertical_count et diagonal_count
-            unsigned score = MAX(horizontal_count, MAX(vertical_count, diagonal_count));
+            unsigned short score = MAX(horizontal_count, MAX(vertical_count, diagonal_count));
             moves_scores[col] = score;
             undo_move(grid, col);
         }
     }
 }
 
-void display_scores(unsigned moves_scores[7]) {
+void display_scores(unsigned short moves_scores[7]) {
     printf("Computed scores: \t");
-    for (unsigned index = 0; index < 7; index++)
+    for (unsigned short index = 0; index < 7; index++)
         printf("%u ", moves_scores[index]);
     printf("\n");
 }
 
-unsigned get_max_score_column(unsigned moves_scores[7]) {
-    unsigned max_score = 0, max_score_column = 0;
+unsigned short get_max_score_column(unsigned short moves_scores[7]) {
+    unsigned short max_score = 0, max_score_column = 0;
 
-    for (unsigned col = 0; col < 7; col++) {
+    for (unsigned short col = 0; col < 7; col++) {
         if (moves_scores[col] > max_score) {
             max_score = moves_scores[col];
             max_score_column = col;
@@ -130,15 +130,15 @@ unsigned get_max_score_column(unsigned moves_scores[7]) {
     return max_score_column;
 }
 
-unsigned play(unsigned grid[6][7]) {
+unsigned short play(unsigned short grid[6][7]) {
     /*
     Retourne la colonne ayant le score maximal
     */
-    unsigned winning_move = check_winning_move(grid);
+    unsigned short winning_move = check_winning_move(grid);
     if (winning_move == NO_WINNING_MOVE) {
         compute_column_scores(grid);
         // display_scores(moves_scores);
-        unsigned max_score_column = get_max_score_column(moves_scores);
+        unsigned short max_score_column = get_max_score_column(moves_scores);
         if (play_token(grid, max_score_column, COMPUTER_TOKEN_COLOR) == COLUMN_FULL)
             return COLUMN_NUMBER_ERROR;
     }
